+++ src/Server/Application/Camino.Application.Contracts/AppServices/Products/Dtos/ProductCategoryRequest.cs
@@ -2,11 +2,11 @@
 {
     public class ProductCategoryRequest
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public long UpdatedById { get; set; }
         public long CreatedById { get; set; }
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
     }
 }
