+++ src/Server/Application/Camino.Application.Contracts/AppServices/Products/Dtos/ProductCategoryResult.cs
@@ -2,7 +2,7 @@
 {
     public class ProductCategoryResult
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public DateTimeOffset UpdatedDate { get; set; }
@@ -11,7 +11,7 @@
         public DateTimeOffset CreatedDate { get; set; }
         public long CreatedById { get; set; }
         public string CreatedBy { get; set; }
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
         public string ParentCategoryName { get; set; }
         public int StatusId { get; set; }
         public ProductCategoryResult ParentCategory { get; set; }
