+++ src/Server/Application/Camino.Application.Contracts/AppServices/Products/IProductCategoryAppService.cs
@@ -5,15 +5,15 @@ namespace Camino.Application.Contracts.AppServices.Products
     public interface IProductCategoryAppService
     {
         Task<bool> ActiveAsync(ProductCategoryRequest request);
-        Task<int> CreateAsync(ProductCategoryRequest request);
+        Task<long> CreateAsync(ProductCategoryRequest request);
         Task<bool> DeactivateAsync(ProductCategoryRequest request);
-        Task<bool> DeleteAsync(int id);
-        Task<ProductCategoryResult> FindAsync(int id);
+        Task<bool> DeleteAsync(long id);
+        Task<ProductCategoryResult> FindAsync(long id);
         Task<ProductCategoryResult> FindByNameAsync(string name);
         Task<BasePageList<ProductCategoryResult>> GetAsync(ProductCategoryFilter filter);
         Task PopulateDetailsAsync(IList<ProductCategoryResult> productCategories);
-        Task<IList<ProductCategoryResult>> SearchAsync(BaseFilter filter, int[] currentIds);
-        Task<IList<ProductCategoryResult>> SearchParentsAsync(BaseFilter filter, int[] currentIds);
+        Task<IList<ProductCategoryResult>> SearchAsync(BaseFilter filter, long[] currentIds);
+        Task<IList<ProductCategoryResult>> SearchParentsAsync(BaseFilter filter, long[] currentIds);
         IList<SelectOption> SearchStatus(IdRequestFilter<int?> filter, string search = "");
         Task<bool> UpdateAsync(ProductCategoryRequest request);
     }
