+++ src/Server/Core/Camino.Core.Domains/Farms/Repositories/IFarmTypeRepository.cs
@@ -4,7 +4,7 @@
     {
         Task<FarmType> FindAsync(long id);
         Task<FarmType> FindByNameAsync(string name);
-        Task<int> CreateAsync(FarmType farmType);
+        Task<long> CreateAsync(FarmType farmType);
         Task<bool> UpdateAsync(FarmType farmType);
         Task<bool> DeleteAsync(int id);
     }
