+++ src/Server/Core/Camino.Core.Domains/Products/ProductCategoryRelation.cs
@@ -4,7 +4,7 @@
     {
         public long Id { get; set; }
         public long ProductId { get; set; }
-        public int ProductCategoryId { get; set; }
+        public long ProductCategoryId { get; set; }
         public virtual Product Product { get; set; }
         public virtual ProductCategory ProductCategory { get; set; }
     }
