+++ src/Server/Core/Camino.Core.Domains/Products/Repositories/IProductCategoryRepository.cs
@@ -4,11 +4,11 @@ namespace Camino.Core.Contracts.Repositories.Products
 {
     public interface IProductCategoryRepository
     {
-        Task<ProductCategory> FindAsync(int id);
-        Task<int> CreateAsync(ProductCategory category);
+        Task<ProductCategory> FindAsync(long id);
+        Task<long> CreateAsync(ProductCategory category);
         Task<bool> UpdateAsync(ProductCategory category);
         Task<ProductCategory> FindByNameAsync(string name);
-        Task<bool> DeleteAsync(int id);
-        Task<bool> HasProductsAsync(int categoryId);
+        Task<bool> DeleteAsync(long id);
+        Task<bool> HasProductsAsync(long categoryId);
     }
 }
