+++ src/Server/Modules/Module.Api.Product/Models/ProductCategorySelectFilterModel.cs
@@ -5,6 +5,6 @@ namespace Module.Api.Product.Models
     public class ProductCategorySelectFilterModel : BaseSelectFilterModel
     {
         public bool? IsParentOnly { get; set; }
-        public int[] CurrentIds { get; set; }
+        public long[] CurrentIds { get; set; }
     }
 }
